requirements.txt
crawl4ai
pinecone
langchain
python-dotenv

